@@ -1,5 +1,4 @@
 import 'package:equatable/equatable.dart';
-import 'package:flutter/cupertino.dart';
 
 abstract class LoginEvent extends Equatable {
   LoginEvent() : super();
@@ -56,7 +55,7 @@ class LoginWithCredentialsPressed extends LoginEvent {
 class ForgotPasswordPressed extends LoginEvent {
   final String email;
 
-  ForgotPasswordPressed({@required this.email})
+  ForgotPasswordPressed({required this.email})
       : super();
 
   @override
