@@ -1,6 +1,5 @@
 import 'package:GTUBT/models/ticket.dart';
 import 'package:equatable/equatable.dart';
-import 'package:flutter/widgets.dart';
 
 abstract class TicketPageState extends Equatable {}
 
@@ -17,7 +16,7 @@ class TicketsLoading extends TicketPageState {
 class TicketsLoaded extends TicketPageState {
   final List<Ticket> tickets;
 
-  TicketsLoaded({@required this.tickets});
+  TicketsLoaded({required this.tickets});
 
   @override
   List<Object> get props => [];
