@@ -3,11 +3,11 @@ import 'package:GTUBT/exceptions/base.dart';
 class EventException extends BaseException {
   final String message;
   
-  EventException([String message])
+  EventException([String? message])
       : message = message == null ? defaultMessage : message,
         super();
 
-  factory EventException.message(String message) {
+  factory EventException.message(String? message) {
     return EventException(message);
   }
 
