@@ -3,11 +3,11 @@ import 'package:GTUBT/exceptions/base.dart';
 class NewsException extends BaseException {
   final String message;
   
-  NewsException([String message])
+  NewsException([String? message])
       : message = message == null ? defaultMessage : message,
         super();
 
-  factory NewsException.message(String message) {
+  factory NewsException.message(String? message) {
     return NewsException(message);
   }
 
