@@ -13,7 +13,7 @@ class EventService extends BaseService {
     return _eventService;
   }
 
-  Future<List<Event>> getAll() async {
+  Future<List<Event?>?> getAll() async {
     String url = '$baseUrl/$endpointPrefix/$servicePath';
 
     final response = await GET('$url');
@@ -33,7 +33,7 @@ class EventService extends BaseService {
     }
   }
 
-  Future<Event> delete(String id) async {
+  Future<Event?> delete(String id) async {
     String url = '$baseUrl/$endpointPrefix/$servicePath/$id';
 
     final response = await DELETE('$url');
@@ -53,7 +53,7 @@ class EventService extends BaseService {
     }
   }
 
-  Future<Event> patch(Event event) async {
+  Future<Event?> patch(Event event) async {
     var id = event.id;
     String url = '$baseUrl/$endpointPrefix/$servicePath/$id';
     try {
@@ -82,7 +82,7 @@ class EventService extends BaseService {
     }
   }
 
-  Future<Event> get(String id) async {
+  Future<Event?> get(String id) async {
     String url = '$baseUrl/$endpointPrefix/$servicePath/$id';
 
     final response = await GET('$url');
