@@ -4,9 +4,9 @@ import 'package:GTUBT/models/user.dart';
 
 void main() {
   test('Get User with email', () async {
-    User user = await UserService().get("yasir.nacak@gmail.com");
+    User? user = await UserService().get("yasir.nacak@gmail.com");
     expect(user == null, false);
-    expect(user.name, "Yasir");
+    expect(user!.name, "Yasir");
   });
 
   test('Post User', () async {
@@ -20,34 +20,34 @@ void main() {
     user.year = 2016;
     user.profilePhoto = "google.com";
 
-    User newUser = await UserService().post(user);
+    User? newUser = await UserService().post(user);
     expect(newUser == null, false);
-    expect(newUser.studentId, user.studentId);
+    expect(newUser!.studentId, user.studentId);
   });
 
   test('Update User', () async {
-    User user = await UserService().get("ahmtergn5@gmail.com");
+    User? user = await UserService().get("ahmtergn5@gmail.com");
     expect(user == null, false);
-    expect(user.name, "Ahmet");
+    expect(user!.name, "Ahmet");
 
     user.phone = "5464351277";
 
-    User updatedUser = await UserService().patch(user);
+    User? updatedUser = await UserService().patch(user);
     expect(updatedUser == null, false);
-    expect(updatedUser.phone, "5464351277");
+    expect(updatedUser!.phone, "5464351277");
 
     user = await UserService().get("ahmtergn5@gmail.com");
     expect(user == null, false);
-    expect(user.phone, "0546 435 12 77");
+    expect(user!.phone, "0546 435 12 77");
   });
 
   test('Delete User', () async {
-    User user = await UserService().get("ahmtergn5@gmail.com");
+    User? user = await UserService().get("ahmtergn5@gmail.com");
     
     expect(user == null, false);
-    expect(user.name, "Ahmet");
+    expect(user!.name, "Ahmet");
 
-    bool isDeleted = await UserService().delete(user.email);
+    bool isDeleted = await UserService().delete(user.email!);
     expect(isDeleted, true);
   });
 }
